requirements.txt
wandb
timm
tqdm
pandas
numpy
scikit-learn
torchmetrics
matplotlib
seaborn
wandb